+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Version: 1.0.17
 * add shimmer integration for bloom effects
 * add man-made compass pages
 * add byproduct limiting
+* add most of the secondary colors to material items
 * improve textures
 * improve zh-cn Localization
 * fix steam boiler can accept steam